+++ backend/accounts/urls.py
@@ -1,4 +1,4 @@
-"""backend URL Configuration
+"""accounts URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.0/topics/http/urls/
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from backend import views
+from accounts import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
