+++ backend/accounts/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for backend project.
+WSGI config for accounts project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'backend.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'accounts.settings')
 
 application = get_wsgi_application()
